loopy: colour in connected regions
Give sufficiently large regions that are known to be connected a single
colour.

This is a very barebones implementation, I'm not actually very happy
with the colour choices or with how they're chosen.  Choosing something
stable or something that assigns adjacent regions different colours
would be nice, but it's harder to implement.